requirements.txt
asgiref==3.8.1
certifi==2025.1.31
cfgv==3.4.0
distlib==0.3.9
Django==5.1.4
djangorestframework==3.15.2
djangorestframework_simplejwt==5.4.0
filelock==3.16.1
identify==2.6.5
nodeenv==1.9.1
pillow==11.1.0
platformdirs==4.3.6
pre_commit==4.0.1
PyJWT==2.10.1
python-dotenv==1.0.1
PyYAML==6.0.2
sqlparse==0.5.3
virtualenv==20.28.1
